Внедрил зависимости с помощью Java Config
