+++ pom.xml
@@ -28,6 +28,12 @@
             <version>2.8.6</version>
             <scope>compile</scope>
         </dependency>
+        <dependency>
+            <groupId>org.springframework</groupId>
+            <artifactId>spring-context</artifactId>
+            <version>5.3.23</version>
+        </dependency>
+
     </dependencies>
 
 </project>
