+++ src/main/java/ru/netology/config/JavaConfig.java
@@ -1,0 +1,26 @@
+package ru.netology.config;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import ru.netology.controller.PostController;
+import ru.netology.repository.PostRepository;
+import ru.netology.repository.PostRepositoryStubImpl;
+import ru.netology.service.PostService;
+
+@Configuration
+public class JavaConfig {
+    @Bean
+    public PostRepository postRepository() {
+        return new PostRepositoryStubImpl();
+    }
+
+    @Bean
+    public PostService postService(PostRepository repository) {
+        return new PostService(repository);
+    }
+
+    @Bean
+    public PostController postController(PostService service) {
+        return new PostController(service);
+    }
+}
