+++ src/main/java/ru/netology/service/PostService.java
@@ -3,6 +3,7 @@ package ru.netology.service;
 import ru.netology.exception.NotFoundException;
 import ru.netology.model.Post;
 import ru.netology.repository.PostRepository;
+import ru.netology.repository.PostRepositoryStubImpl;
 
 import java.util.List;
